vty: add a unioned index_int field
The problem here is that the index pointer field got reused in some contexts
to store int64_t values, which is almost OK if you have 64-bit pointers, but
which can break things if you don't (-Wpointer-to-int-cast). Fixing it is
non-trivial, because this field is used in a lot of places to also store
proper pointers, so what's done here is an anonymous union of this field with
an index_int of int64_t type and some associated fixes.

Anonymous unions are a part of GNU C99 (which this codebase is targetting) and
also a part of proper standard C11, so it should be fine to use them.